backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
cloud-sql-python-connector>=1.4.0
google-cloud-aiplatform>=1.46.0
google-cloud-logging>=3.9.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.25
pydantic>=2.6.0
python-multipart>=0.0.6
psycopg2-binary>=2.9.9
redis>=5.0.1
httpx>=0.26.0
feedparser>=6.0.11
vertexai>=0.10.0
